backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt==3.2.0
passlib[bcrypt]
python-multipart
httpx
celery
redis
apscheduler
python-dotenv
alembic
